prediction-service/requirements.txt
grpcio>=1.63.0
grpcio-tools>=1.63.0
protobuf>=5.26.1
python-dotenv==1.0.1
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
psycopg2-binary>=2.9.9
schedule>=1.2.1
